split to another component
