+++ Client/src/views/Sidebar/Sidebar.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import { Icon, Menu, Sidebar } from 'semantic-ui-react';
+import '../../Semantic-UI-CSS-master/semantic.min.css';
+import MenuItem from './MenuItem';
+
+const SidebarApp = () => {
+    return(
+        <Sidebar
+            as={Menu}
+            animation='uncover'
+            direction={"left"}
+            icon='labeled'
+            inverted
+            vertical
+            visible='true'
+            width='thin'
+        >
+            <MenuItem iconName='database' iconText='Backup and restore'/>
+            <MenuItem iconName='dochub' iconText='Documentation'/>
+            <MenuItem iconName='docker' iconText='Docker CLI'/>
+        </Sidebar>
+    );
+};
+
+export default SidebarApp
